+++ CHANGELOG.md
@@ -5,8 +5,14 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## v1.0.0 - 2020-11-13
+### Added
+- Added mockserver and tests.
+- Added `LocalAddr() net.Addr` and `RemoteAddr() net.Addr` functions that returns local and remote network addresses.
+
 ## v0.1.0 - 2020-10-22
 ### Added
 - Initial implementation.
 
-[Unreleased]: https://github.com/gorcon/websocket/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/gorcon/websocket/compare/v1.0.0...HEAD
+[v1.0.0]: https://github.com/gorcon/rcon/compare/v0.1.0...v1.0.0